Define new BPF object that separates out the maps and types in headers
