+++ pkg/ebpf/c/maps.bpf.c
@@ -1,0 +1,1 @@
+#include maps.bpf.h
