+++ pkg/ebpf/c/tracee.bpf.c
@@ -78,6 +78,9 @@
 #include <bpf/bpf_tracing.h>
 #include <bpf/bpf_endian.h>
 
+#include "types.bpf.h"
+#include "maps.bpf.h"
+
 char LICENSE[] SEC("license") = "GPL";
 #ifndef CORE
 int KERNEL_VERSION SEC("version") = LINUX_VERSION_CODE;
@@ -648,39 +651,6 @@ enum kconfig_key_e
         })
 #endif
 
-// EBPF MAP MACROS ---------------------------------------------------------------------------------
-
-#define BPF_MAP(_name, _type, _key_type, _value_type, _max_entries)                                \
-    struct {                                                                                       \
-        __uint(type, _type);                                                                       \
-        __uint(max_entries, _max_entries);                                                         \
-        __type(key, _key_type);                                                                    \
-        __type(value, _value_type);                                                                \
-    } _name SEC(".maps");
-
-#define BPF_HASH(_name, _key_type, _value_type, _max_entries)                                      \
-    BPF_MAP(_name, BPF_MAP_TYPE_HASH, _key_type, _value_type, _max_entries)
-
-#define BPF_LRU_HASH(_name, _key_type, _value_type, _max_entries)                                  \
-    BPF_MAP(_name, BPF_MAP_TYPE_LRU_HASH, _key_type, _value_type, _max_entries)
-
-#define BPF_ARRAY(_name, _value_type, _max_entries)                                                \
-    BPF_MAP(_name, BPF_MAP_TYPE_ARRAY, u32, _value_type, _max_entries)
-
-#define BPF_PERCPU_ARRAY(_name, _value_type, _max_entries)                                         \
-    BPF_MAP(_name, BPF_MAP_TYPE_PERCPU_ARRAY, u32, _value_type, _max_entries)
-
-#define BPF_PROG_ARRAY(_name, _max_entries)                                                        \
-    BPF_MAP(_name, BPF_MAP_TYPE_PROG_ARRAY, u32, u32, _max_entries)
-
-#define BPF_PERF_OUTPUT(_name, _max_entries)                                                       \
-    BPF_MAP(_name, BPF_MAP_TYPE_PERF_EVENT_ARRAY, int, __u32, _max_entries)
-
-// stack traces: the value is 1 big byte array of the stack addresses
-typedef __u64 stack_trace_t[MAX_STACK_DEPTH];
-#define BPF_STACK_TRACE(_name, _max_entries)                                                       \
-    BPF_MAP(_name, BPF_MAP_TYPE_STACK_TRACE, u32, stack_trace_t, _max_entries)
-
 #ifndef CORE
     #ifdef RHEL_RELEASE_CODE
         #if (RHEL_RELEASE_CODE >= RHEL_RELEASE_VERSION(8, 0))
@@ -691,285 +661,6 @@ typedef __u64 stack_trace_t[MAX_STACK_DEPTH];
         #error Minimal required kernel version is 4.18
     #endif
 #endif
-
-// INTERNAL STRUCTS --------------------------------------------------------------------------------
-
-// todo: check network context after this change
-typedef struct task_context {
-    u64 start_time; // thread's start time
-    u64 cgroup_id;
-    u32 pid;       // PID as in the userspace term
-    u32 tid;       // TID as in the userspace term
-    u32 ppid;      // Parent PID as in the userspace term
-    u32 host_pid;  // PID in host pid namespace
-    u32 host_tid;  // TID in host pid namespace
-    u32 host_ppid; // Parent PID in host pid namespace
-    u32 uid;
-    u32 mnt_id;
-    u32 pid_id;
-    char comm[TASK_COMM_LEN];
-    char uts_name[TASK_COMM_LEN];
-    u32 flags;
-} task_context_t;
-
-typedef struct event_context {
-    u64 ts; // Timestamp
-    task_context_t task;
-    u32 eventid;
-    u32 padding;
-    s64 retval;
-    u32 stack_id;
-    u16 processor_id; // The ID of the processor which processed the event
-    u8 argnum;
-} event_context_t;
-
-typedef struct args {
-    unsigned long args[6];
-} args_t;
-
-typedef struct syscall_data {
-    uint id;           // Current syscall id
-    args_t args;       // Syscall arguments
-    unsigned long ts;  // Timestamp of syscall entry
-    unsigned long ret; // Syscall ret val. May be used by syscall exit tail calls.
-} syscall_data_t;
-
-typedef struct fd_arg_task {
-    u32 pid;
-    u32 tid;
-    int fd;
-} fd_arg_task_t;
-
-typedef struct fd_arg_path {
-    char path[MAX_CACHED_PATH_SIZE];
-} fd_arg_path_t;
-
-typedef struct task_info {
-    task_context_t context;
-    syscall_data_t syscall_data;
-    bool syscall_traced;  // indicates that syscall_data is valid
-    bool recompute_scope; // recompute should_trace (new task/context changed/policy changed)
-    bool new_task;        // set if this task was started after tracee. Used with new_pid filter
-    bool follow;          // set if this task was traced before. Used with the follow filter
-    int should_trace;     // last decision of should_trace()
-    u8 container_state;   // the state of the container the task resides in
-} task_info_t;
-
-typedef struct bin_args {
-    u8 type;
-    u8 metadata[SEND_META_SIZE];
-    char *ptr;
-    loff_t start_off;
-    unsigned int full_size;
-    u8 iov_idx;
-    u8 iov_len;
-    struct iovec *vec;
-} bin_args_t;
-
-typedef struct simple_buf {
-    u8 buf[MAX_PERCPU_BUFSIZE];
-} buf_t;
-
-typedef struct path_filter {
-    char path[MAX_PATH_PREF_SIZE];
-} path_filter_t;
-
-typedef struct string_filter {
-    char str[MAX_STR_FILTER_SIZE];
-} string_filter_t;
-
-typedef struct ksym_name {
-    char str[MAX_KSYM_NAME_SIZE];
-} ksym_name_t;
-
-typedef struct config_entry {
-    u32 tracee_pid;
-    u32 options;
-    u32 filters;
-    u32 cgroup_v1_hid;
-    u64 uid_max;
-    u64 uid_min;
-    u64 pid_max;
-    u64 pid_min;
-    u64 mnt_ns_max;
-    u64 mnt_ns_min;
-    u64 pid_ns_max;
-    u64 pid_ns_min;
-    u8 events_to_submit[128]; // use 8*128 bits to describe up to 1024 events
-} config_entry_t;
-
-typedef struct event_data {
-    event_context_t context;
-    struct task_struct *task;
-    task_info_t *task_info;
-    void *ctx;
-    config_entry_t *config;
-    buf_t *submit_p;
-    u32 buf_off;
-} event_data_t;
-
-// For a good summary about capabilities, see https://lwn.net/Articles/636533/
-typedef struct slim_cred {
-    uid_t uid;           // real UID of the task
-    gid_t gid;           // real GID of the task
-    uid_t suid;          // saved UID of the task
-    gid_t sgid;          // saved GID of the task
-    uid_t euid;          // effective UID of the task
-    gid_t egid;          // effective GID of the task
-    uid_t fsuid;         // UID for VFS ops
-    gid_t fsgid;         // GID for VFS ops
-    u32 user_ns;         // User Namespace of the event
-    u32 securebits;      // SUID-less security management
-    u64 cap_inheritable; // caps our children can inherit
-    u64 cap_permitted;   // caps we're permitted
-    u64 cap_effective;   // caps we can actually use
-    u64 cap_bset;        // capability bounding set
-    u64 cap_ambient;     // Ambient capability set
-} slim_cred_t;
-
-typedef struct network_connection_v4 {
-    u32 local_address;
-    u16 local_port;
-    u32 remote_address;
-    u16 remote_port;
-} net_conn_v4_t;
-
-typedef struct network_connection_v6 {
-    struct in6_addr local_address;
-    u16 local_port;
-    struct in6_addr remote_address;
-    u16 remote_port;
-    u32 flowinfo;
-    u32 scope_id;
-} net_conn_v6_t;
-
-typedef struct net_id {
-    struct in6_addr address;
-    u16 port;
-    u16 protocol;
-} net_id_t;
-
-typedef struct net_packet {
-    uint64_t ts;
-    u32 event_id;
-    u32 host_tid;
-    char comm[TASK_COMM_LEN];
-    u32 len;
-    u32 ifindex;
-    struct in6_addr src_addr, dst_addr;
-    __be16 src_port, dst_port;
-    u8 protocol;
-} net_packet_t;
-
-typedef struct net_ctx {
-    u32 host_tid;
-    char comm[TASK_COMM_LEN];
-} net_ctx_t;
-
-typedef struct net_ctx_ext {
-    u32 host_tid;
-    char comm[TASK_COMM_LEN];
-    __be16 local_port;
-} net_ctx_ext_t;
-
-// version is not size limited - save only first 32 bytes.
-// srcversion is not size limited - modpost calculates srcversion with size: 25.
-#define MODULE_VERSION_MAX_LENGTH    32
-#define MODULE_SRCVERSION_MAX_LENGTH 25
-
-typedef struct kmod_data {
-    char name[MODULE_NAME_LEN];
-    char version[MODULE_VERSION_MAX_LENGTH];
-    char srcversion[MODULE_SRCVERSION_MAX_LENGTH];
-    u64 prev;
-    u64 next;
-} kmod_data_t;
-
-typedef struct file_info {
-    char pathname[MAX_CACHED_PATH_SIZE];
-    dev_t device;
-    unsigned long inode;
-    u64 ctime;
-} file_info_t;
-
-typedef struct bpf_attach_key {
-    u32 host_tid;
-    u32 prog_id;
-} bpf_attach_key_t;
-
-typedef struct bpf_attach {
-    enum bpf_write_user_e write_user;
-} bpf_attach_t;
-
-// KERNEL STRUCTS ----------------------------------------------------------------------------------
-
-#ifndef CORE
-struct mnt_namespace {
-    atomic_t count;
-    struct ns_common ns;
-    // ...
-};
-
-struct mount {
-    struct hlist_node mnt_hash;
-    struct mount *mnt_parent;
-    struct dentry *mnt_mountpoint;
-    struct vfsmount mnt;
-    // ...
-};
-
-    #define get_type_size(x)            sizeof(x)
-    #define get_node_addr(array, index) &array[index]
-
-#endif
-
-// EBPF MAPS DECLARATIONS --------------------------------------------------------------------------
-
-// clang-format off
-BPF_HASH(kconfig_map, u32, u32, 10240);                            // kernel config variables
-BPF_HASH(interpreter_map, u32, file_info_t, 10240);                // interpreter file used for each process
-BPF_HASH(containers_map, u32, u8, 10240);                          // map cgroup id to container status {EXISTED, CREATED, STARTED}
-BPF_HASH(args_map, u64, args_t, 1024);                             // persist args between function entry and return
-BPF_HASH(uid_filter, u32, u32, 256);                               // filter events by UID, for specific UIDs either by == or !=
-BPF_HASH(pid_filter, u32, u32, 256);                               // filter events by PID
-BPF_HASH(mnt_ns_filter, u64, u32, 256);                            // filter events by mount namespace id
-BPF_HASH(pid_ns_filter, u64, u32, 256);                            // filter events by pid namespace id
-BPF_HASH(uts_ns_filter, string_filter_t, u32, 256);                // filter events by uts namespace name
-BPF_HASH(comm_filter, string_filter_t, u32, 256);                  // filter events by command name
-BPF_HASH(cgroup_id_filter, u32, u32, 256);                         // filter events by cgroup id
-BPF_HASH(bin_args_map, u64, bin_args_t, 256);                      // persist args for send_bin funtion
-BPF_HASH(sys_32_to_64_map, u32, u32, 1024);                        // map 32bit to 64bit syscalls
-BPF_HASH(params_types_map, u32, u64, 1024);                        // encoded parameters types for event
-BPF_HASH(process_tree_map, u32, u32, 10240);                       // filter events by the ancestry of the traced process
-BPF_LRU_HASH(task_info_map, u32, task_info_t, 10240);              // holds data for every task
-BPF_HASH(network_config, u32, int, 1024);                          // holds the network config for each iface
-BPF_HASH(ksymbols_map, ksym_name_t, u64, 1024);                    // holds the addresses of some kernel symbols
-BPF_HASH(syscalls_to_check_map, int, u64, 256);                    // syscalls to discover
-BPF_LRU_HASH(sock_ctx_map, u64, net_ctx_ext_t, 10240);             // socket address to process context
-BPF_LRU_HASH(network_map, net_id_t, net_ctx_t, 10240);             // network identifier to process context
-BPF_ARRAY(config_map, config_entry_t, 1);                          // various configurations
-BPF_ARRAY(file_filter, path_filter_t, 3);                          // filter vfs_write events
-BPF_PERCPU_ARRAY(bufs, buf_t, MAX_BUFFERS);                        // percpu global buffer variables
-BPF_PROG_ARRAY(prog_array, MAX_TAIL_CALL);                         // store programs for tail calls
-BPF_PROG_ARRAY(prog_array_tp, MAX_TAIL_CALL);                      // store programs for tail calls
-BPF_PROG_ARRAY(sys_enter_tails, MAX_EVENT_ID);                     // store syscall specific programs for tail calls from sys_enter
-BPF_PROG_ARRAY(sys_exit_tails, MAX_EVENT_ID);                      // store syscall specific programs for tail calls from sys_exit
-BPF_PROG_ARRAY(sys_enter_submit_tail, MAX_EVENT_ID);               // store program for submitting syscalls from sys_enter
-BPF_PROG_ARRAY(sys_exit_submit_tail, MAX_EVENT_ID);                // store program for submitting syscalls from sys_exit
-BPF_PROG_ARRAY(sys_enter_init_tail, MAX_EVENT_ID);                 // store program for performing syscall tracking logic in sys_enter
-BPF_PROG_ARRAY(sys_exit_init_tail, MAX_EVENT_ID);                  // store program for performing syscall tracking logic in sys_exits
-BPF_STACK_TRACE(stack_addresses, MAX_STACK_ADDRESSES);             // store stack traces
-BPF_HASH(module_init_map, u32, kmod_data_t, 256);                  // holds module information between
-BPF_LRU_HASH(fd_arg_path_map, fd_arg_task_t, fd_arg_path_t, 1024); // store fds paths by task
-BPF_LRU_HASH(bpf_attach_map, u32, bpf_attach_t, 1024);             // holds bpf prog info
-BPF_LRU_HASH(bpf_attach_tmp_map, u32, bpf_attach_t, 1024);         // temporarily hold bpf_attach_t
-// clang-format on
-
-// EBPF PERF BUFFERS -------------------------------------------------------------------------------
-
-BPF_PERF_OUTPUT(events, 1024);      // events submission
-BPF_PERF_OUTPUT(file_writes, 1024); // file writes events submission
-BPF_PERF_OUTPUT(net_events, 1024);  // network events submission
 
 // HELPERS: DEVICES --------------------------------------------------------------------------------
